+++ character/views/character.py
@@ -33,7 +33,7 @@ class CharacterCreateView(LoginRequiredMixin, CreateView):
     template_name = "character/character_create.html"
 
     def get_success_url(self):
-        return reverse("equipment-select", args=(self.object.id,))
+        return reverse("skills-select", args=(self.object.id,))
 
     def _apply_racial_traits(self, character, racial_trait):
         character.adult_age = racial_trait.adult_age
@@ -44,7 +44,7 @@ class CharacterCreateView(LoginRequiredMixin, CreateView):
         # Need to save before setting many-to-many relationships
         character.save()
         character.languages.set(racial_trait.languages.all())
-        character.abilities.set(racial_trait.abilities.all())
+        character.senses.set(racial_trait.senses.all())
 
     def _apply_ability_score_increases(self, character, racial_trait):
         ability_score_increases = AbilityScoreIncrease.objects.filter(